Updating code so that the gui is separate from the background logic.
